pyproject.toml
[project]
name = "crypto-bot-assistant"
version = "0.1.0"
description = "Telegram bot for crypto analysis using Bybit API"
requires-python = ">=3.13"

dependencies = [
    # Telegram API
    "python-telegram-bot>=21.10",  # Последняя версия

    # Crypto exchange support
    "ccxt>=4.4.64",  # Последняя версия

    # Data analysis and plotting
    "numpy>=1.26.4",
    "pandas>=2.2.3",
    "matplotlib>=3.8.3",
    "ta>=0.11.0",  # Technical analysis indicators

    # Async and requests
    "httpx>=0.27.0",  # Для асинхронных запросов

    # Environment and validation
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.4",

    # Logging
    "loguru>=0.7.2",

    # Web framework (на будущее для расширения)
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",

    # Database support
    "psycopg2-binary>=2.9.10",

    # Security and authentication
    "oauthlib>=3.2.2",

    # Email validation
    "email-validator>=2.2.0",

    # WSGI сервер (для развертывания на продакшене)
    "gunicorn>=23.0.0"
]

[tool.setuptools]
packages = ["crypto_bot_assistant"]
